BUG: too many file opens (no more connections accepted by the server)
This happens when there is no data to be transmitted in one or multiple
channels. seedlink raw client time out, close its side of the connection
and reconnect. Unfortunately the raw_server doesn't notice the closed
connection and keep the socket open on its side. Eventually the system
limit of max files open is reached.